requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
weasyprint==60.2
redis==5.0.1
celery==5.3.6
pytest==7.4.3
requests==2.31.0
